Major Map Overhaul
Map Class:

Revamped Map class significantly. It now stores a 2D array of Nodes, each of which holds one Location and all of its reachable neighbors (any Location that's next to it and doesn't share a wall with it). The findPath() function has been made private, and a new getPath() function that checks if a path has already been found between two Locations and only searches for a new path if it hasn't. Changed both pathfinding functions to return lists of DirTypes instead of lists of Locations. Added getter functions for the list of all Coin Locations and the list of Diamond Locations.

CoinBotStrategy Class:
Modified code to work with the changes in the Map class.

ScoutBotStrategy1 Class:
Moved deadEnd() code to the Map, which now automatically marks dead ends.

GhostBotStrategy Class:
New stubbed class. Will write this one since it will be pretty simple.