Added GhostBotStrategy
Un-stubbed GhostBotStrategy. GhostBot heads directly for the closest Diamond. Added GhostBot and VisionBot to OurSearchingTeam.

Realized that we can't directly store anything that Prof.Lepinski gives us. The Location objects given to us are different every time, even for the same locations in the Maze. May need to make a new class for them or add to the Node code.