Refined coinBotStrategy
added an unknown weighting function that is supposed to take in each path, and create a best-case scenario for all unknown squares.
This impacts path length additively, and for every unknown square, the number gets larger and larger. This can be changed back to the initial method that we have by using 'setUnknownPref(true);'
this isn't totally finished, I don't think